New Project
Today I reorganized my projects, and I made a simple web page that allows a user to input the details of a book, and see the book getting added to the webpage if the information filled is correct.
For this project I used the recent acquired knowledge in creating and management of classes, objects and literals.